[titusan] reorganizing the signup file and breaking things out into more maintainable components
