+++ Pathread/app/screens/SignUp/constants/interests.ts
@@ -1,0 +1,8 @@
+export const INTERESTS = [
+  'Art & Design', 'Gaming', 'Hiking', 'Exercising', 'Music',
+  'Photography', 'Cooking', 'Travel', 'Reading', 'Movies',
+  'Sports', 'Dancing', 'Writing', 'Technology', 'Nature',
+  'Fashion', 'Meditation', 'Gardening', 'Crafting', 'Astronomy',
+  'History', 'Science', 'Languages', 'Volunteering', 'Collecting',
+  'Theater', 'Comedy', 'Fitness', 'Yoga', 'Puzzles'
+] as const;
