+++ Pathread/app/screens/SignUp/styles/signUp.styles.ts
@@ -1,0 +1,297 @@
+import { StyleSheet } from 'react-native';
+
+export const signUpStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#000000',
+  },
+  bgGradient: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+  },
+  content: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: 24,
+    paddingTop: 20,
+    paddingBottom: 20,
+  },
+  titleContainer: {
+    alignItems: 'center',
+    marginBottom: 40,
+  },
+  title: {
+    color: '#ffffff',
+    fontSize: 32,
+    fontWeight: '600',
+    letterSpacing: 1,
+    textTransform: 'lowercase',
+    marginBottom: 8,
+  },
+  whoAreYou: {
+    color: '#ffffff',
+    fontSize: 36,
+    fontWeight: '700',
+    textAlign: 'center',
+  },
+  birthdayQuestion: {
+    color: '#ffffff',
+    fontSize: 36,
+    fontWeight: '700',
+    textAlign: 'center',
+  },
+  subtitleContainer: {
+    alignItems: 'center',
+    marginBottom: 40,
+    paddingHorizontal: 20,
+  },
+  subtitle: {
+    color: '#9aa0a6',
+    fontSize: 16,
+    textAlign: 'center',
+    lineHeight: 22,
+  },
+  formContainer: {
+    width: '100%',
+    marginBottom: 24,
+  },
+  input: {
+    backgroundColor: 'rgba(15,15,18,0.8)',
+    borderWidth: 1,
+    borderColor: 'rgba(120,130,150,0.3)',
+    borderRadius: 16,
+    paddingHorizontal: 20,
+    paddingVertical: 18,
+    color: '#ffffff',
+    fontSize: 16,
+    textAlign: 'center',
+  },
+  inputRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    gap: 12,
+  },
+  halfInput: {
+    flex: 1,
+  },
+  dropdownButton: {
+    backgroundColor: 'rgba(15,15,18,0.8)',
+    borderWidth: 1,
+    borderColor: 'rgba(120,130,150,0.3)',
+    borderRadius: 16,
+    paddingHorizontal: 20,
+    paddingVertical: 18,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    height: 56,
+  },
+  dropdownText: {
+    color: '#ffffff',
+    fontSize: 16,
+  },
+  placeholderText: {
+    color: '#7a7f85',
+  },
+  modalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.8)',
+    justifyContent: 'flex-end',
+    alignItems: 'center',
+  },
+  modalContent: {
+    backgroundColor: 'rgba(15,15,18,0.98)',
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+    width: '100%',
+    maxHeight: '60%',
+    borderWidth: 1,
+    borderColor: 'rgba(120,130,150,0.3)',
+    borderBottomWidth: 0,
+  },
+  modalTitle: {
+    color: '#ffffff',
+    fontSize: 20,
+    fontWeight: '700',
+    textAlign: 'center',
+    paddingVertical: 20,
+    borderBottomWidth: 1,
+    borderBottomColor: 'rgba(120,130,150,0.2)',
+  },
+  modalScrollView: {
+    maxHeight: 300,
+  },
+  modalOption: {
+    paddingVertical: 18,
+    paddingHorizontal: 20,
+    borderBottomWidth: 1,
+    borderBottomColor: 'rgba(120,130,150,0.1)',
+  },
+  selectedOption: {
+    backgroundColor: 'rgba(74, 144, 226, 0.2)',
+  },
+  modalOptionText: {
+    color: '#ffffff',
+    fontSize: 16,
+    textAlign: 'center',
+    fontWeight: '500',
+  },
+  selectedOptionText: {
+    color: '#4A90E2',
+    fontWeight: '600',
+  },
+  modalCancel: {
+    marginTop: 10,
+    marginHorizontal: 20,
+    marginBottom: 20,
+    paddingVertical: 16,
+    paddingHorizontal: 20,
+    backgroundColor: 'rgba(120,130,150,0.2)',
+    borderRadius: 12,
+  },
+  modalCancelText: {
+    color: '#9aa0a6',
+    fontSize: 16,
+    textAlign: 'center',
+    fontWeight: '600',
+  },
+  // Results step styles
+  youAreContainer: {
+    alignItems: 'center',
+    marginBottom: 15,
+  },
+  youAreText: {
+    color: '#9aa0a6',
+    fontSize: 20,
+    fontWeight: '500',
+    textAlign: 'center',
+    letterSpacing: 1,
+  },
+  iconContainer: {
+    alignItems: 'center',
+    marginBottom: 15,
+  },
+  icon: {
+    fontSize: 100,
+    textAlign: 'center',
+    textShadowColor: '#ffa500',
+    textShadowOffset: { width: 0, height: 0 },
+    textShadowRadius: 15,
+  },
+  resultsTitleContainer: {
+    alignItems: 'center',
+    marginBottom: 25,
+  },
+  resultsTitle: {
+    color: '#ffa500',
+    fontSize: 32,
+    fontWeight: '700',
+    textAlign: 'center',
+    letterSpacing: 1,
+    textShadowColor: '#ffa500',
+    textShadowOffset: { width: 0, height: 0 },
+    textShadowRadius: 8,
+  },
+  descriptionContainer: {
+    width: '100%',
+    marginBottom: 40,
+    paddingHorizontal: 20,
+  },
+  descriptionText: {
+    color: '#9aa0a6',
+    fontSize: 16,
+    lineHeight: 22,
+    textAlign: 'center',
+    fontWeight: '400',
+  },
+  // Interests step styles
+  interestsTitleContainer: {
+    alignItems: 'center',
+    marginBottom: 12,
+  },
+  interestsTitle: {
+    color: '#ffffff',
+    fontSize: 26,
+    fontWeight: '700',
+    textAlign: 'center',
+  },
+  interestsSubtitleContainer: {
+    alignItems: 'center',
+    marginBottom: 24,
+    paddingHorizontal: 20,
+  },
+  interestsSubtitle: {
+    color: '#9aa0a6',
+    fontSize: 15,
+    textAlign: 'center',
+    lineHeight: 20,
+  },
+  interestsGridContainer: {
+    width: '100%',
+    flex: 1,
+    marginBottom: 20,
+  },
+  interestsScrollView: {
+    flex: 1,
+  },
+  interestsScrollContent: {
+    paddingBottom: 10,
+  },
+  interestsGrid: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    justifyContent: 'space-between',
+    gap: 10,
+  },
+  interestTag: {
+    backgroundColor: 'rgba(15,15,18,0.8)',
+    borderWidth: 1,
+    borderColor: 'rgba(120,130,150,0.3)',
+    borderRadius: 18,
+    paddingHorizontal: 14,
+    paddingVertical: 10,
+    minWidth: '30%',
+    alignItems: 'center',
+    marginBottom: 8,
+  },
+  selectedInterestTag: {
+    backgroundColor: '#4A90E2',
+    borderWidth: 0,
+    overflow: 'hidden',
+  },
+  interestText: {
+    color: '#ffffff',
+    fontSize: 13,
+    fontWeight: '500',
+    textAlign: 'center',
+  },
+  selectedInterestText: {
+    color: '#ffffff',
+    fontWeight: '600',
+  },
+  disabledButton: {
+    opacity: 0.6,
+  },
+  buttonContainer: {
+    width: '100%',
+    paddingTop: 10,
+  },
+  button: {
+    borderRadius: 16,
+    overflow: 'hidden',
+  },
+  buttonGradient: {
+    paddingVertical: 18,
+    alignItems: 'center',
+    borderRadius: 16,
+  },
+  buttonText: {
+    color: '#ffffff',
+    fontWeight: '700',
+    fontSize: 18,
+  },
+});
